refactor(mcp): extract McpCapabilityInference out of McpToolBridge
McpToolBridge had grown to 408 lines, of which ~200 lines were
MCP-method-name pattern matching, JSON-field lookup, and disambiguation
logic for move/copy semantics. The class was simultaneously:

  1. A `Tool` adapter for the MCP protocol (its actual job)
  2. A method-name classifier (verb regexes + camelCase normalization)
  3. A field-name lookup with case/underscore normalization
  4. A capability builder with sensitivity-aware disambiguation

Pure structural refactor: pull (2)(3)(4) into a focused package-private
helper class. McpToolBridge.toCapability is now 3 lines and the bridge
is back to ~160 lines of "MCP adapter for Tool".

  Before:  McpToolBridge.java     408 lines
  After:   McpToolBridge.java     162 lines (-60%)
           McpCapabilityInference  293 lines (new)

Behavior is identical. All 43 tests in McpToolBridgeTest pass
unchanged - the existing transitive coverage proves the integration.
Direct unit tests for McpCapabilityInference can land in a follow-up
if we want package-internal coverage too.
